requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==3.3.2
dill==0.3.8
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.29.3
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
multiprocess==0.70.16
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
propcache==0.3.0
pyarrow==19.0.1
pyperclip==1.9.0
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
# Editable install with no version control (swebench-evaluator==0.1.0)
-e /Users/matei/Working/rbc/swebench-eval
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
